Added some Flutter dependencies
LOL all I did was make some yaml changes and Flutter made all the rest of this. Love you Flutter.

What's happening in my code? Who knows? If the computer does half the work here without my input, shouldn't that just be in the interpereter or compiler and not in the code?

Who designed this stuff?